Add hoover auto cleaning acceptance test
- assume that our clint wishes to add a new feature to the already existing functionality. Our precious robotic hover should be able to clean its dirt container after whe work.
In this case, first, we should create an acceptance test that should check if a container is empty after a room has been cleaned.
We don't have neither logic nor unit tests covering this functionality yet.

Also, in most cases, Cucumber-based tests live in a separate project (like e2e tests), because they are focused on testing the whole system, not just separate  micro services.

In our case, we have included them in this package in the sake of simplicity.